forge-framework/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
python-dotenv==1.0.1
web3==7.14.1
httpx==0.27.2
redis==5.0.8
pytest==9.0.2
black==24.10.0
paho-mqtt==2.1.0
python-multipart==0.0.9
